Technomancer Skill Level Calculator 1.2.7
Instead of Save and Load system, a Coroutine that takes an array of entries is more appropriate. This system is not complete. And the manual log entry and save/load system is not completely deprecated. This new conveyor system is written with the intention of a deviation in program design.